refactor: Change the path of 'Point' related files to 'Geometry'
