+++ Sources/Core/Geometry/TriangleMeshToSDF.cpp
@@ -8,10 +8,10 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
+#include <Core/Geometry/Point3.hpp>
 #include <Core/Geometry/TriangleMesh3.hpp>
 #include <Core/Geometry/TriangleMeshToSDF.hpp>
 #include <Core/Grid/ScalarGrid3.hpp>
-#include <Core/Point/Point3.hpp>
 #include <Core/Size/Size3.hpp>
 #include <Core/Vector/Vector3.hpp>
